+++ adash_tests/tests/adash_tests-suite.adb
@@ -7,6 +7,7 @@ with Adash_Tests.Execution_Cases;
 with Adash_Tests.Interactive_Cases;
 with Adash_Tests.Language_Cases;
 with Adash_Tests.Lexer_Cases;
+with Adash_Tests.Pattern_Cases;
 with Adash_Tests.Message_Cases;
 with Adash_Tests.Parser_Cases;
 with Adash_Tests.Persistence_Cases;
@@ -32,6 +33,7 @@ package body Adash_Tests.Suite is
    Source_Case     : aliased Adash_Tests.Source_Cases.Case_Type;
    Language_Case   : aliased Adash_Tests.Language_Cases.Case_Type;
    Lexer_Case      : aliased Adash_Tests.Lexer_Cases.Case_Type;
+   Pattern_Case    : aliased Adash_Tests.Pattern_Cases.Case_Type;
    Parser_Case     : aliased Adash_Tests.Parser_Cases.Case_Type;
    Semantics_Case  : aliased Adash_Tests.Semantics_Cases.Case_Type;
    Evaluation_Case : aliased Adash_Tests.Evaluation_Cases.Case_Type;
@@ -63,6 +65,7 @@ package body Adash_Tests.Suite is
       AUnit.Test_Suites.Add_Test (Result'Access, Configuration_Case'Access);
       AUnit.Test_Suites.Add_Test (Result'Access, Language_Case'Access);
       AUnit.Test_Suites.Add_Test (Result'Access, Lexer_Case'Access);
+      AUnit.Test_Suites.Add_Test (Result'Access, Pattern_Case'Access);
       AUnit.Test_Suites.Add_Test (Result'Access, Parser_Case'Access);
       AUnit.Test_Suites.Add_Test (Result'Access, Semantics_Case'Access);
       AUnit.Test_Suites.Add_Test (Result'Access, Evaluation_Case'Access);